moved cli_tests as unit tests to args.rs
